New workbook for nested for loops
In module 2.3.2, it stated to create a new workbook.  I created it and named it the same name as the new worksheet.  At the end of the module it stated to upload green_stocks.xlsm which I did not modify for this module and I did not want to overwrite it with this code so I just added this as an additional file.

It is possible that later, I could have used github to retrieve the previous version but I am not sure if that was the intension. If so, I will go back later to replace green_stocks with this file.