New subroutine to analyze all stocks
New subroutine AllStocksAnalysis. 

This subroutine uses nested loops to populate a section of a spreadsheet using variables for the rows/cols. 

After it populates the section it clears it out using the Range method for the active worksheet and the row/col variables that specify where the section is located.