New Buttons and functions added
Adds the following to AllStocksAnalysis and green_stocks workbook:

- New Run Analysis for All Stocks button: runs calls the macro to run all stocks from the 2018 worksheet.
- New Clear Worksheet button: runs the macro to all the user to clear the entire 2018 worksheet.
- New call to formatting macro in AllStocksAnalysis macro so when the user clicks the button from the worksheet the formatting is included.
- New call to inputBox to retrieve the sheet year to run the analysis based on the user's input. 
NOTE: The call to the new subroutine added does not work because the yearValue is not being set in the correct routine.  A global variable could be used but I am not sure if I am doing this correctly. In the exercise, it said this new macro would have the Range lines that are in the Analysis macro so I am not sure if they should be in the new macro as well.

For now, I am calling the InputBox function directly where I need it so this works and I can move on. I will look into this later.