Deliverable 1 Complete
These files contain the code to complete Deliverable 1 for Challenge 2

Note that the xlsm file has both the Excel data and the code and the vbs file has the code as well. This is because there was an issue with VS Code. Once it was configured to work for python I could not get it to work in time to work on the VBA file and run/debug it properly so I switched back to the  Excel Developer tools to complete the code.