Complete code for Deliverable 1
VBA_Challenge.xlsm: contains both the code and the spreadsheet. 

VBA_Challenge.vbs: code was copied from xlsm.

NOTE: the xlsm file contains the code because I had an issue with VS Code once it was configured to work with python and I did not have time to figure out how to get it to run and debug VBA code correctly in time for this challenge.