New subroutine to format data
This version adds a new subroutine to format the heading and data separately from the main code which sets the values.

This makes the code simpler to read and debug in the future. It also separates the functionality so new formatting can easily be added without affecting the data.

Since it only adds 1 iteration through the data vs. an exponential number it should not affect performance.