Correct names for png files.
Named according to challenge requirements.

Since these files do not include the runtime results that are used in the readme file they are included in git for the challenge. The files in the readme file are named to indicate which code was used for the run, the year it was run on.